+++ src/prog2int/Models/CodigoBarras.java
@@ -24,7 +24,7 @@ public class CodigoBarras extends Base {
      * Usado por PersonaDAO al mapear ResultSet.
      * El domicilio se asigna posteriormente con setDomicilio().
      */
-    public CodigoBarras(int id, String tipoCB,String valor, Date fecha, String observaciones) {
+    public CodigoBarras(long id, String tipoCB,String valor, Date fecha, String observaciones) {
         super(id, false);
         this.tipoCB = TipoCB.valueOf(tipoCB.toUpperCase());
         this.valor = valor;
